deepfake_image_detection/requirements.txt
flask
torch
torchvision
tqdm
scikit-learn
pillow
